[SPOT-280][REFACTOR] Organize Story Domain
