+++ src/main/java/com/example/spot/refactor/story/domain/aggregate/LikedStudyComment.java
@@ -1,14 +1,8 @@
-package com.example.spot.refactor.study.domain.aggregate.studypost;
+package com.example.spot.refactor.story.domain.aggregate;
 
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.common.entity.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
@@ -20,16 +14,18 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikedStudyPost extends BaseEntity {
+public class LikedStudyComment extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(nullable = false, columnDefinition = "BIT DEFAULT 1")
+    private Boolean isLiked;
 
     @Setter
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "study_post_id", nullable = false)
-    private StudyPost studyPost;
+    @JoinColumn(name = "study_post_comment_id", nullable = false)
+    private StudyPostComment studyPostComment;
 
     @Setter
     @ManyToOne(fetch = FetchType.LAZY)
