+++ src/main/java/com/example/spot/refactor/story/domain/aggregate/StudyPostComment.java
@@ -1,4 +1,4 @@
-package com.example.spot.refactor.study.domain.aggregate.studypost;
+package com.example.spot.refactor.story.domain.aggregate;
 
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.common.entity.BaseEntity;
