+++ src/main/java/com/example/spot/refactor/story/domain/enums/StudyPostCategory.java
@@ -1,4 +1,4 @@
-package com.example.spot.refactor.study.domain.enums;
+package com.example.spot.refactor.story.domain.enums;
 
 public enum StudyPostCategory {
     WELCOME, INFO_SHARING, STUDY_REVIEW, FREE_TALK, QNA
