+++ src/main/java/com/example/spot/refactor/story/domain/repository/StudyPostRepositoryCustom.java
@@ -1,6 +1,7 @@
-package com.example.spot.refactor.study.domain.aggregate.studypost;
+package com.example.spot.refactor.story.domain.repository;
 
-import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategory;
 
 import org.springframework.data.domain.Pageable;
