+++ src/main/java/com/example/spot/refactor/story/domain/validation/validator/ExistLikedStudyCommentValidator.java
@@ -1,8 +1,8 @@
-package com.example.spot.refactor.study.domain.validation.validator;
+package com.example.spot.refactor.story.domain.validation.validator;
 
 import com.example.spot.refactor.common.api.code.status.ErrorStatus;
-import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyCommentRepository;
-import com.example.spot.refactor.study.domain.validation.annotation.ExistLikedStudyComment;
+import com.example.spot.refactor.story.domain.repository.LikedStudyCommentRepository;
+import com.example.spot.refactor.story.domain.validation.annotation.ExistLikedStudyComment;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
