+++ src/main/java/com/example/spot/refactor/story/domain/validation/validator/ExistStudyPostValidator.java
@@ -1,8 +1,8 @@
-package com.example.spot.refactor.study.domain.validation.validator;
+package com.example.spot.refactor.story.domain.validation.validator;
 
 import com.example.spot.refactor.common.api.code.status.ErrorStatus;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
-import com.example.spot.refactor.study.domain.validation.annotation.ExistStudyPost;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
+import com.example.spot.refactor.story.domain.validation.annotation.ExistStudyPost;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
