+++ src/main/java/com/example/spot/refactor/study/application/MemberStudyCommandServiceImpl.java
@@ -21,14 +21,14 @@ import com.example.spot.refactor.member.domain.enums.Status;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVoteParticipant;
 import com.example.spot.refactor.study.domain.aggregate.Study;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostReport;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPostReport;
 import com.example.spot.legacy.repository.MemberReportRepository;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.legacy.repository.NotificationRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostReportRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostReportRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
 import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.study.presentation.dto.request.ScheduleRequestDTO;
