+++ src/main/java/com/example/spot/refactor/study/application/MemberStudyQueryServiceImpl.java
@@ -15,12 +15,12 @@ import com.example.spot.refactor.study.domain.aggregate.studyvote.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.schedule.domain.enums.StudySchedulePeriod;
 import com.example.spot.refactor.study.domain.aggregate.Study;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
 import com.example.spot.refactor.todo.domain.StudyToDo;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVote;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
 import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.study.presentation.dto.response.ScheduleResponseDTO;
