+++ src/main/java/com/example/spot/refactor/study/application/StudyPostCommandServiceImpl.java
@@ -5,11 +5,11 @@ import com.example.spot.refactor.common.api.exception.handler.MemberHandler;
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.legacy.domain.Notification;
+import com.example.spot.refactor.story.domain.aggregate.*;
+import com.example.spot.refactor.story.domain.repository.*;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
-import com.example.spot.refactor.study.domain.aggregate.studypost.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.legacy.domain.enums.NotifyType;
-import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyPost;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
