+++ src/main/java/com/example/spot/refactor/study/application/StudyPostQueryServiceImpl.java
@@ -5,16 +5,16 @@ import com.example.spot.refactor.common.api.exception.handler.MemberHandler;
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategoryQuery;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategory;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategoryQuery;
 import com.example.spot.refactor.study.domain.aggregate.Study;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostComment;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPostComment;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyPostRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostCommentRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.LikedStudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostCommentRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
 import com.example.spot.refactor.common.security.utils.SecurityUtils;
 import com.example.spot.refactor.study.presentation.dto.response.StudyPostCommentResponseDTO;
