+++ src/main/java/com/example/spot/refactor/study/presentation/controller/StudyPostController.java
@@ -2,13 +2,13 @@ package com.example.spot.refactor.study.presentation.controller;
 
 import com.example.spot.refactor.common.api.ApiResponse;
 import com.example.spot.refactor.common.api.code.status.SuccessStatus;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategoryQuery;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategoryQuery;
 import com.example.spot.legacy.service.s3.S3ImageService;
 import com.example.spot.refactor.study.application.StudyPostCommandService;
 import com.example.spot.refactor.study.application.StudyPostQueryService;
 import com.example.spot.refactor.study.domain.validation.annotation.ExistStudy;
-import com.example.spot.refactor.study.domain.validation.annotation.ExistStudyPost;
-import com.example.spot.refactor.study.domain.validation.annotation.ExistStudyPostComment;
+import com.example.spot.refactor.story.domain.validation.annotation.ExistStudyPost;
+import com.example.spot.refactor.story.domain.validation.annotation.ExistStudyPostComment;
 import com.example.spot.refactor.study.presentation.dto.request.StudyPostCommentRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyPostRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.response.StudyPostCommentResponseDTO;
