+++ src/main/java/com/example/spot/refactor/study/presentation/dto/response/StudyPostResDTO.java
@@ -1,9 +1,9 @@
 package com.example.spot.refactor.study.presentation.dto.response;
 
 import com.example.spot.refactor.member.domain.Member;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostImage;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategory;
+import com.example.spot.refactor.story.domain.aggregate.StudyPostImage;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
 import lombok.*;
 
 import java.time.LocalDateTime;
