+++ src/test/java/com/example/spot/service/study/studymember/StudyMemberQueryServiceTest.java
@@ -6,13 +6,13 @@ import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.schedule.domain.StudySchedule;
 import com.example.spot.refactor.todo.domain.StudyToDo;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategory;
 import com.example.spot.refactor.study.domain.aggregate.Study;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
-import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
 import com.example.spot.refactor.todo.domain.StudyToDoRepository;
 import com.example.spot.refactor.common.security.utils.SecurityUtils;
 import com.example.spot.refactor.study.application.MemberStudyQueryServiceImpl;
