+++ src/test/java/com/example/spot/service/study/studypost/StudyPostQueryServiceTest.java
@@ -2,13 +2,18 @@ package com.example.spot.service.study.studypost;
 
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
+import com.example.spot.refactor.story.domain.aggregate.LikedStudyComment;
+import com.example.spot.refactor.story.domain.aggregate.LikedStudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPost;
+import com.example.spot.refactor.story.domain.aggregate.StudyPostComment;
+import com.example.spot.refactor.story.domain.repository.LikedStudyPostRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostCommentRepository;
+import com.example.spot.refactor.story.domain.repository.StudyPostRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
-import com.example.spot.refactor.study.domain.aggregate.studypost.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.member.domain.enums.Gender;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
-import com.example.spot.refactor.study.domain.enums.StudyPostCategoryQuery;
-import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyComment;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategory;
+import com.example.spot.refactor.story.domain.enums.StudyPostCategoryQuery;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
